Moves tab related files from js to app folder, split components to a single file per component
Resolves #8347

Auditors: @bsclifton @cezaraugusto

Test Plan:
- test should be green
