+++ app/renderer/components/tabs/tab.js
@@ -3,32 +3,44 @@
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 const React = require('react')
-
-const ImmutableComponent = require('./immutableComponent')
 const {StyleSheet, css} = require('aphrodite')
-
-const windowActions = require('../actions/windowActions')
-const appActions = require('../actions/appActions')
-const locale = require('../l10n')
-const dragTypes = require('../constants/dragTypes')
-const messages = require('../constants/messages')
-const cx = require('../lib/classSet')
-const {getTextColorForBackground} = require('../lib/color')
-const {isIntermediateAboutPage} = require('../lib/appUrlUtil')
-
-const contextMenus = require('../contextMenus')
-const dnd = require('../dnd')
-const windowStore = require('../stores/windowStore')
 const ipc = require('electron').ipcRenderer
-const throttle = require('../lib/throttle')
 
-const styles = require('../../app/renderer/components/styles/tab')
-const {Favicon, AudioTabIcon, NewSessionIcon,
-      PrivateIcon, TabTitle, CloseTabIcon} = require('../../app/renderer/components/tabContent')
-const {getTabBreakpoint, tabUpdateFrameRate} = require('../../app/renderer/lib/tabUtil')
-const {isWindows} = require('../../app/common/lib/platformUtil')
-const {getCurrentWindowId} = require('../../app/renderer/currentWindow')
-const {frameOptsFromFrame} = require('../state/frameStateUtil')
+// Components
+const ImmutableComponent = require('../../../../js/components/immutableComponent')
+const Favicon = require('./content/favIcon')
+const AudioTabIcon = require('./content/audioTabIcon')
+const NewSessionIcon = require('./content/newSessionIcon')
+const PrivateIcon = require('./content/privateIcon')
+const TabTitle = require('./content/tabTitle')
+const CloseTabIcon = require('./content/closeTabIcon')
+
+// Actions
+const windowActions = require('../../../../js/actions/windowActions')
+const appActions = require('../../../../js/actions/appActions')
+
+// Store
+const windowStore = require('../../../../js/stores/windowStore')
+const {frameOptsFromFrame} = require('../../../../js/state/frameStateUtil')
+
+// Constants
+const dragTypes = require('../../../../js/constants/dragTypes')
+const messages = require('../../../../js/constants/messages')
+
+// Styles
+const styles = require('../styles/tab')
+
+// Utils
+const locale = require('../../../../js/l10n')
+const cx = require('../../../../js/lib/classSet')
+const {getTextColorForBackground} = require('../../../../js/lib/color')
+const {isIntermediateAboutPage} = require('../../../../js/lib/appUrlUtil')
+const contextMenus = require('../../../../js/contextMenus')
+const dnd = require('../../../../js/dnd')
+const throttle = require('../../../../js/lib/throttle')
+const {getTabBreakpoint, tabUpdateFrameRate} = require('../../lib/tabUtil')
+const {isWindows} = require('../../../common/lib/platformUtil')
+const {getCurrentWindowId} = require('../../currentWindow')
 
 class Tab extends ImmutableComponent {
   constructor () {
@@ -378,8 +390,8 @@ class Tab extends ImmutableComponent {
 }
 
 const paymentsEnabled = () => {
-  const getSetting = require('../settings').getSetting
-  const settings = require('../constants/settings')
+  const getSetting = require('../../../../js/settings').getSetting
+  const settings = require('../../../../js/constants/settings')
   return getSetting(settings.PAYMENTS_ENABLED)
 }
